Update Software_Carbon_Intensity_Specification.md
This is how the conversation with @SaraEmilyBergman and @Sealjay evolved over the telemetry vs. benchmark discussions. We realised we were talking about measuring in the real-world (telemetry) and measuring in more of a lab environment (benchmark). The important point wasn't telemetry vs benchmark which was just implementation details, it was making it clear that we cover both. Modelling is important since a huge subset of the audience for SCI won't be able to measure using real-world data and will have to model, e.g. most open source projects and anyone who values the privacy of their users.