Remove “marginal” due to lack of scientific consensus. Add hourly granularity.
There is currently no scientific consensus on which decision-making signal (marginal, average, potential profiles, etc.) performs best. On top of that, there is no ground truth for marginal - it is impossible to verify whether the marginal value is true or not. As long as the debate is ongoing, it is risky to prescribe marginal emissions as the only valid signal in the SCI and thus we should be open to other metrics. Any metric(s) included in the SCI need to be verifiable and lead to real emissions reductions and with it decarbonise the electricity system.

Additionally, this specification focuses on temporal granularity, which is why "hourly" was added.

Sources:
https://www.cell.com/iscience/fulltext/S2589-0042(22)00185-7?_returnURL=https%3A%2F%2Flinkinghub.elsevier.com%2Fretrieve%2Fpii%2FS2589004222001857%3Fshowall%3Dtrue

https://www.pnas.org/doi/10.1073/pnas.2211624119

https://media.rff.org/documents/Report_22-08.pdf

https://cebi.org/wp-content/uploads/2022/11/Guide-to-Sourcing-Marginal-Emissions-Factor-Data.pdf