Remove marginal and with it the explanation. Add hourly granularity.
The marginal emission rate is defined as the carbon intensity of the marginal generator. While the marginal emissions rate does indeed represent a value with units gCO2/MWh, it does not necessarily represent the change in emissions with a change in load. There are two main scenarios: 1. if the load is very small, no additional generation will be deployed and the additional load is reflected in a slight change of grid frequency. No additional emissions are emitted as a result of the added load. 2. If the load is large, the marginal emissions rate is not guaranteed to represent the change in emissions, since an additional marginal generator (with a different MER) could be called upon to service the large additional load. The marginal emission factor therefore does not generally apply to consumers with different intervention sizes.

Marginal is removed as there is currently no scientific consensus on which signal is the best for short-term decision-making. As long as the debate is ongoing, it is risky to prescribe marginal emissions as the only valid signal in the SCI.

Refer to the previous commit regarding hourly granularity.