Remove “marginal” due to lack of scientific consensus, to allow for different metrics.
As highlighted in previous commits, prescribing marginal would pose a risk since past and currently ongoing research indicates doubts whether MERs are able to properly quantify impacts of interventions. Moreover, there is currently no ground truth on MER as it is practically impossible to verify the marginal generator. We advocate to not prescribe a specific carbon intensity metric until there is scientific consensus on it being superior than others. Any metric(s) included in the SCI need to be verifiable and lead to real emissions reductions and with it decarbonise the electricity system.

Sources:
https://www.cell.com/iscience/fulltext/S2589-0042(22)00185-7?_returnURL=https%3A%2F%2Flinkinghub.elsevier.com%2Fretrieve%2Fpii%2FS2589004222001857%3Fshowall%3Dtrue

https://www.pnas.org/doi/10.1073/pnas.2211624119

https://media.rff.org/documents/Report_22-08.pdf

https://cebi.org/wp-content/uploads/2022/11/Guide-to-Sourcing-Marginal-Emissions-Factor-Data.pdf

See here for further discussions:
https://github.com/Green-Software-Foundation/sci/discussions/351#discussioncomment-4936341